+++ examples/linearlite/src/pages/Issue/Comments.tsx
@@ -5,7 +5,7 @@ import Avatar from '../../components/Avatar'
 import { formatDate } from '../../utils/date'
 import { showWarning } from '../../utils/notification'
 import { Issue } from '../../types'
-import { useStore, useTemporaryQuery } from '@livestore/react'
+import { useStore, useScopedQuery } from '@livestore/react'
 import { querySQL, sql } from '@livestore/livestore'
 import { nanoid } from 'nanoid'
 import { mutations, tables } from '../../domain/schema'
@@ -19,7 +19,7 @@ export interface CommentsProps {
 function Comments({ issue }: CommentsProps) {
   // TODO move this into LiveStore
   const [newCommentBody, setNewCommentBody] = useState<string>('')
-  const comments = useTemporaryQuery(
+  const comments = useScopedQuery(
     () =>
       querySQL(() => sql`SELECT * FROM comment WHERE issueId = '${issue.id}' ORDER BY created ASC`, {
         schema: Schema.Array(tables.comment.schema),