+++ packages/@livestore/react/src/experimental/components/LiveList.tsx
@@ -3,7 +3,7 @@ import { computed } from '@livestore/livestore'
 import React from 'react'
 
 import { useQuery } from '../../useQuery.js'
-import { useTemporaryQuery } from '../../useTemporaryQuery.js'
+import { useScopedQuery } from '../../useScopedQuery.js'
 
 /*
 TODO:
@@ -33,7 +33,7 @@ export const LiveList = <TItem,>({ items$, renderItem, getKey }: LiveListProps<T
   React.useEffect(() => setHasMounted(true), [])
 
   const keysCb = React.useCallback(() => computed((get) => get(items$).map(getKey)), [getKey, items$])
-  const keys = useTemporaryQuery(keysCb, 'fixed')
+  const keys = useScopedQuery(keysCb, 'fixed')
   const arr = React.useMemo(
     () =>
       keys.map(